Set your geobounds cache dir

This function will store your cache_dir path on your local machine and would load it for future sessions. Type gb_detect_cache_dir() to find your cached path.

Alternatively, you can store the cache_dir manually with the following options:

Run Sys.setenv(GEOBOUNDS_CACHE_DIR = "cache_dir"). You would need to run this command on each session (Similar to install = FALSE).

Write this line on your .Renviron file: GEOBOUNDS_CACHE_DIR = "value_for_cache_dir" (same behavior than install = TRUE). This would store your cache_dir permanently.

Usage

gb_set_cache_dir(cache_dir, overwrite = FALSE, install = FALSE, quiet = FALSE)

Arguments

cache_dir

A path to a cache directory. On missing value the function would store the cached files on a temporary dir (See base::tempdir()).

overwrite

Logical. If this is set to TRUE, it will overwrite an existing cache_dir.

install

Logical. If TRUE, will install the key in your local machine for use in future sessions. Defaults to FALSE. If cache_dir is FALSE this parameter is set to FALSE automatically.

quiet

Logical, on FALSE it displays information of the call. Useful for debugging, default is no messages quiet = TRUE.

Value

An (invisible) character with the path to your cache_dir.

See also

Examples

# Don't run this! It would modify your current state
# \dontrun{
gb_set_cache_dir(quiet = FALSE)
#> ℹ Using a temporary cache directory. Set `cache_dir` to a value for store permanently
#> ✔ geobounds cache dir is C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpeyh68P/geobounds.
# }

gb_detect_cache_dir()
#> ℹ C:\Users\RUNNER~1\AppData\Local\Temp\Rtmpeyh68P/geobounds
#> [1] "C:\\Users\\RUNNER~1\\AppData\\Local\\Temp\\Rtmpeyh68P/geobounds"